refactor: Move 'getExtension' function to utils folder - #1
